Clean up logging
